Move admin navigation Playwright test to dedicated file
Agent-Logs-Url: https://github.com/OrchardCMS/OrchardCore/sessions/6a2afe3f-da95-4bf0-858d-a65d8c1556a3
